Update the sidebar entries based on the root URI.
The sidebar generation code has been made more flexible so that it can be
run more than once, instead of assuming it will only ever be run once.

Various sidebar entries are now dependent on the type of root. Search, for
example, only shows up with a file:-based root.

When deciding whether a volume can be added, the root URI is checked to make
sure the volume resides within that root.