Keep the list of custom shortcuts in a list and rebuild it on root_uri change.
The old custom shortcut code assumed that the entry would be added to the
sidebar and never removed programatically within that instance of the file
chooser. That, clearly, doesn't work when dynamically changing the root URIs.

We now maintain a list of added custom shortcuts and add/remove these based
on the root URI whenever the root changes.