Add support for multiple roots.
This introduces support for multiple roots on the file chooser. Instead
of rooting to a single URI, there can now be several roots, which will be
taken into account when selecting files, typing custom file paths, or showing
parts of the UI.

This makes it possible to, for example, limit the file chooser to the user's
home directory and memory stick.