Filter the path bar and auto-completion results.
This changes the path bar to only show those folders accessible within
the root URI. If a particular folder has been set as the root URI, it will
appear as the top-most button in the path bar, much like the File System
does.

This also hides the scroll buttons in the Path Bar when no buttons are set
(in the case of an invalid root URI) and hides the left scroll button if the
first shown folder is the very first folder in the path bar.