Fix error case in filesystem model
We should not unref the model here, it might not even exist anymore.
Instead check if it exists and only use it if it does.
The unref was leftover from a previous fix in
ba9f53397ff0b86ad56166c7fc91b4d499f7aefc.

Spotted by Matthias Clasen in
https://bugzilla.gnome.org/show_bug.cgi?id=614099