menu: Fix wrong comparison causing navigation region to not get removed
The navigation region that got created for right-facing submenus when
moving the cursor up would have two negative values which caused the
check to go wrong.

I refactored the code to make that check into a separate function as it
makes the check not only harder to get wrong, but also the code easier
to read.

http://bugzilla.gnome.org/show_bug.cgi?id=620863