Remove usage of meta_rect()
This was just of the form

  MetaRectangle r = meta_rect (x, y, w, h);

But it was now superfluous in a couple of cases where we could
just use the GdkRectangle that we already had (i.e. without
having to create a new MetaRectangle), and in another
case where all fields were set to zero.

And lo and behold, we actually don't need any meta_rectangle_*()
functions in the theme code.