GtkTreeview now watches the allocation of the parent scrolled window at allocation time in order to detect possible feedback loops.
To avoid locking up the desktop some feedback protection is needed,
this patch is not the best but as far as I can see it does the job.
Ideally we should be looking into a GtkScrollable interface that
takes this height-for-width possibility into account.

The problem being that some height-for-width calculations at
allocation time are asynchronous and at some sizes cause the
height-for-width logic to compete with parent allocations
which change due to scrollbars toggling visiblity.

Specific considerations
  a.) The child treeview should not have knowlage of its parents
      allocation or the size of the scrollbars (this patch cheats
      that idealistic rule).
  b.) The parent scrolled window cannot safely determine if a child
      wants its size renegotiated based on a content change or based
      on a previous allocation