Added gtk_spread_table_get_child_line
Gets the line index in which a child would be positioned
if the table were to be allocated size in the opposing
orientation of the table.

For instance, if the table is oriented vertically,
this function will return the child's column if
the table were to be allocated size width.